requirements.txt
fastapi
uvicorn[standard]
apify-client
httpx
python-dotenv
pydantic-settings
weasyprint
openai==0.28.1
anyio
asyncio
markdown
python-multipart    
pillow
transformers
google-generativeai




